requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
pypdf>=5.0.0
groq>=0.9.0
numpy>=1.26.0
pydantic-settings>=2.3.0
pytest>=8.0.0
httpx>=0.27.0
